refactor: organiza a inicialização dos módulos de controllers em arquivos separados
